packer/target/logtrixia/aws-monitoring/requirements.txt
elasticsearch>=7.0.0,<8.0.0
numpy
boto3
flatten_json
python-dotenv